* commit '62ce89a568ab316d85919a9ac0d5a061ac5410f6':
  内存监测
  宫格布局时白板没充分利用的bug
  添加内存泄漏监测库
  白板SDKV1.2.1.0
  补充log
  完成直播举手
  恢复
  增加了直播中接收到个人禁言的处理，优化了全体禁言和个人禁言
  上传图片时删除不必要的代码
  no message
  no message
  一对一时全屏课件老师视频拖动，error.localizedDescription
  一对一时，全屏课件老师视频拖动，error.localizedDescription
  直播举手按钮的位置变化
  V2.7.0.1
  1V1全屏时 老师视频尺寸修改
  SDK V2.1.7
  直播举手，去掉上麦

# Conflicts:
#	YSLive/YSSmallClass/SCMainVC.m
#	YSLive/YSTeacherRole/YSTeacherRoleMainVC.m
#	YSLiveManager/YSLiveManager.m

+++ YSLiveManager/YSLiveSignaling/YSLiveManager+SendTeacherSignaling.m
@@ -585,7 +585,7 @@
 {
     NSInteger classStartTime = self.tClassStartTime;//1584526733
     
-    NSString * msgID = [NSString stringWithFormat:@"RaiseHandStart_%@_%ld",self.room_Id,classStartTime];
+    NSString * msgID = [NSString stringWithFormat:@"RaiseHandStart_%@_%ld",self.room_Id,(long)classStartTime];
     //RaiseHandStart_1931343076_1584526733.000000
     
     BOOL result = [self.roomManager pubMsg:YSSignalingName_RaiseHandStart msgID:msgID toID:YSRoomPubMsgTellAll data:@{@"maxSort":@300,@"subInterval":@2500} save:YES extensionData:@{@"type":@"useSort"} associatedMsgID:nil associatedUserID:nil expires:0 completion:completion];
